+++ src/Builder/Realtime.php
@@ -9,7 +9,7 @@
  * file that was distributed with this source code.
  */
 
-namespace ConsoleTVs\Charts;
+namespace ConsoleTVs\Charts\Builder;
 
 /**
  * This is the realtime class.
